+++ lazypredict/Supervised.py
@@ -11,90 +11,143 @@ import time
 import sklearn
 from sklearn.pipeline import Pipeline
 from sklearn.impute import SimpleImputer, MissingIndicator
-from sklearn.preprocessing import StandardScaler, OneHotEncoder
+from sklearn.preprocessing import StandardScaler, OneHotEncoder, OrdinalEncoder
 from sklearn.compose import ColumnTransformer
 from sklearn.utils.testing import all_estimators
 from sklearn.base import RegressorMixin
 from sklearn.base import ClassifierMixin
-from sklearn.metrics import accuracy_score, balanced_accuracy_score, roc_auc_score, f1_score, r2_score, mean_squared_error
+from sklearn.metrics import (
+    accuracy_score,
+    balanced_accuracy_score,
+    roc_auc_score,
+    f1_score,
+    r2_score,
+    mean_squared_error,
+)
 import warnings
 import xgboost
+
 # import catboost
 import lightgbm
 
 warnings.filterwarnings("ignore")
 pd.set_option("display.precision", 2)
-pd.set_option("display.float_format", lambda x: '%.2f' % x)
+pd.set_option("display.float_format", lambda x: "%.2f" % x)
 
-CLASSIFIERS = [est for est in all_estimators(
-) if issubclass(est[1], ClassifierMixin)]
-REGRESSORS = [est for est in all_estimators(
-) if issubclass(est[1], RegressorMixin)]
+CLASSIFIERS = [est for est in all_estimators() if issubclass(est[1], ClassifierMixin)]
+REGRESSORS = [est for est in all_estimators() if issubclass(est[1], RegressorMixin)]
 
-removed_classifiers = [('ClassifierChain', sklearn.multioutput.ClassifierChain),
- ('ComplementNB', sklearn.naive_bayes.ComplementNB),
- ('GradientBoostingClassifier',
-  sklearn.ensemble.gradient_boosting.GradientBoostingClassifier),
- ('GaussianProcessClassifier',sklearn.gaussian_process.gpc.GaussianProcessClassifier),
- ('HistGradientBoostingClassifier',
-  sklearn.ensemble._hist_gradient_boosting.gradient_boosting.HistGradientBoostingClassifier),
- ('MLPClassifier', sklearn.neural_network.multilayer_perceptron.MLPClassifier),
- ('LogisticRegressionCV', sklearn.linear_model.logistic.LogisticRegressionCV),
- ('MultiOutputClassifier', sklearn.multioutput.MultiOutputClassifier),
- ('MultinomialNB', sklearn.naive_bayes.MultinomialNB),
- ('OneVsOneClassifier', sklearn.multiclass.OneVsOneClassifier),
- ('OneVsRestClassifier', sklearn.multiclass.OneVsRestClassifier),
- ('OutputCodeClassifier', sklearn.multiclass.OutputCodeClassifier),
- ('RadiusNeighborsClassifier',
-  sklearn.neighbors.classification.RadiusNeighborsClassifier),
- ('VotingClassifier', sklearn.ensemble.voting.VotingClassifier)]
+removed_classifiers = [
+    ("CheckingClassifier", sklearn.utils._mocking.CheckingClassifier),
+    ("ClassifierChain", sklearn.multioutput.ClassifierChain),
+    ("ComplementNB", sklearn.naive_bayes.ComplementNB),
+    (
+        "GradientBoostingClassifier",
+        sklearn.ensemble.gradient_boosting.GradientBoostingClassifier,
+    ),
+    (
+        "GaussianProcessClassifier",
+        sklearn.gaussian_process.gpc.GaussianProcessClassifier,
+    ),
+    (
+        "HistGradientBoostingClassifier",
+        sklearn.ensemble._hist_gradient_boosting.gradient_boosting.HistGradientBoostingClassifier,
+    ),
+    ("MLPClassifier", sklearn.neural_network.multilayer_perceptron.MLPClassifier),
+    ("LogisticRegressionCV", sklearn.linear_model.logistic.LogisticRegressionCV),
+    ("MultiOutputClassifier", sklearn.multioutput.MultiOutputClassifier),
+    ("MultinomialNB", sklearn.naive_bayes.MultinomialNB),
+    ("OneVsOneClassifier", sklearn.multiclass.OneVsOneClassifier),
+    ("OneVsRestClassifier", sklearn.multiclass.OneVsRestClassifier),
+    ("OutputCodeClassifier", sklearn.multiclass.OutputCodeClassifier),
+    (
+        "RadiusNeighborsClassifier",
+        sklearn.neighbors.classification.RadiusNeighborsClassifier,
+    ),
+    ("VotingClassifier", sklearn.ensemble.voting.VotingClassifier),
+]
 
-removed_regressors = [('TheilSenRegressor', sklearn.linear_model.theil_sen.TheilSenRegressor),
- ('ARDRegression', sklearn.linear_model.ARDRegression),
- ('CCA', sklearn.cross_decomposition.CCA),
- ('IsotonicRegression', sklearn.isotonic.IsotonicRegression),
- ('MultiOutputRegressor', sklearn.multioutput.MultiOutputRegressor),
- ('MultiTaskElasticNet',
-  sklearn.linear_model.MultiTaskElasticNet),
- ('MultiTaskElasticNetCV',
-  sklearn.linear_model.MultiTaskElasticNetCV),
- ('MultiTaskLasso', sklearn.linear_model.MultiTaskLasso),
- ('MultiTaskLassoCV',
-  sklearn.linear_model.MultiTaskLassoCV),
- ('PLSCanonical', sklearn.cross_decomposition.PLSCanonical),
- ('PLSRegression', sklearn.cross_decomposition.PLSRegression),
- ('RadiusNeighborsRegressor',
-  sklearn.neighbors.RadiusNeighborsRegressor),
- ('RegressorChain', sklearn.multioutput.RegressorChain),
- ('VotingRegressor', sklearn.ensemble.VotingRegressor),
- ('_SigmoidCalibration', sklearn.calibration._SigmoidCalibration)]
+removed_regressors = [
+    ("TheilSenRegressor", sklearn.linear_model.theil_sen.TheilSenRegressor),
+    ("ARDRegression", sklearn.linear_model.ARDRegression),
+    ("CCA", sklearn.cross_decomposition.CCA),
+    ("IsotonicRegression", sklearn.isotonic.IsotonicRegression),
+    ("StackingRegressor",sklearn.ensemble.StackingRegressor),
+    ("MultiOutputRegressor", sklearn.multioutput.MultiOutputRegressor),
+    ("MultiTaskElasticNet", sklearn.linear_model.MultiTaskElasticNet),
+    ("MultiTaskElasticNetCV", sklearn.linear_model.MultiTaskElasticNetCV),
+    ("MultiTaskLasso", sklearn.linear_model.MultiTaskLasso),
+    ("MultiTaskLassoCV", sklearn.linear_model.MultiTaskLassoCV),
+    ("PLSCanonical", sklearn.cross_decomposition.PLSCanonical),
+    ("PLSRegression", sklearn.cross_decomposition.PLSRegression),
+    ("RadiusNeighborsRegressor", sklearn.neighbors.RadiusNeighborsRegressor),
+    ("RegressorChain", sklearn.multioutput.RegressorChain),
+    ("VotingRegressor", sklearn.ensemble.VotingRegressor),
+    ("_SigmoidCalibration", sklearn.calibration._SigmoidCalibration),
+]
 
 for i in removed_regressors:
     REGRESSORS.pop(REGRESSORS.index(i))
-    
+
 for i in removed_classifiers:
     CLASSIFIERS.pop(CLASSIFIERS.index(i))
 
-REGRESSORS.append(('XGBRegressor', xgboost.XGBRegressor))
-REGRESSORS.append(('LGBMRegressor',lightgbm.LGBMRegressor))
+REGRESSORS.append(("XGBRegressor", xgboost.XGBRegressor))
+REGRESSORS.append(("LGBMRegressor", lightgbm.LGBMRegressor))
 # REGRESSORS.append(('CatBoostRegressor',catboost.CatBoostRegressor))
-    
-CLASSIFIERS.append(('XGBClassifier',xgboost.XGBClassifier))
-CLASSIFIERS.append(('LGBMClassifier',lightgbm.LGBMClassifier))
+
+CLASSIFIERS.append(("XGBClassifier", xgboost.XGBClassifier))
+CLASSIFIERS.append(("LGBMClassifier", lightgbm.LGBMClassifier))
 # CLASSIFIERS.append(('CatBoostClassifier',catboost.CatBoostClassifier))
 
-numeric_transformer = Pipeline(steps=[
-    ('imputer', SimpleImputer(strategy='mean')),
-    ('scaler', StandardScaler())
-])
+numeric_transformer = Pipeline(
+    steps=[("imputer", SimpleImputer(strategy="mean")), ("scaler", StandardScaler())]
+)
 
-categorical_transformer = Pipeline(steps=[
-    ('imputer', SimpleImputer(strategy='constant', fill_value='missing')),
-    ('encoding', OneHotEncoder(handle_unknown='ignore', sparse=False))
-])
+categorical_transformer_low = Pipeline(
+    steps=[
+        ("imputer", SimpleImputer(strategy="constant", fill_value="missing")),
+        ("encoding", OneHotEncoder(handle_unknown="ignore", sparse=False)),
+    ]
+)
+
+categorical_transformer_high = Pipeline(
+    steps=[
+        ("imputer", SimpleImputer(strategy="constant", fill_value="missing")),
+        # 'OrdianlEncoder' Raise a ValueError when encounters an unknown value. Check https://github.com/scikit-learn/scikit-learn/pull/13423
+        ("encoding", OrdinalEncoder()),
+    ]
+)
+
+
+# Helper function
+
+
+def get_card_split(df, cols, n=11):
+    """
+    Splits categorical columns into 2 lists based on cardinality (i.e # of unique values)
+    Parameters
+    ----------
+    df : Pandas DataFrame
+        DataFrame from which the cardinality of the columns is calculated.
+    cols : list-like
+        Categorical columns to list
+    n : int, optional (default=11)
+        The value of 'n' will be used to split columns.
+    Returns
+    -------
+    card_low : list-like
+        Columns with cardinality < n
+    card_high : list-like
+        Columns with cardinality >= n
+    """
+    cond = df[cols].nunique() > n
+    card_high = cols[cond]
+    card_low = cols[~cond]
+    return card_low, card_high
+
 
 # Helper class for performing classification
-
 
 class LazyClassifier:
     """
@@ -110,6 +163,8 @@ class LazyClassifier:
         When function is provided, models are evaluated based on the custom evaluation metric provided.
     prediction : bool, optional (default=False)
         When set to True, the predictions of all the models models are returned as dataframe.
+    classifiers : list, optional (default="all")
+        When function is provided, trains the chosen classifier(s).
 
     Examples
     --------
@@ -122,6 +177,7 @@ class LazyClassifier:
     >>> X_train, X_test, y_train, y_test = train_test_split(X, y,test_size=.5,random_state =123)
     >>> clf = LazyClassifier(verbose=0,ignore_warnings=True, custom_metric=None)
     >>> models,predictions = clf.fit(X_train, X_test, y_train, y_test)
+    >>> model_dictionary = clf.provide_models(X_train,X_test,y_train,y_test)
     >>> models
     | Model                          |   Accuracy |   Balanced Accuracy |   ROC AUC |   F1 Score |   Time Taken |
     |:-------------------------------|-----------:|--------------------:|----------:|-----------:|-------------:|
@@ -157,12 +213,22 @@ class LazyClassifier:
     | DummyClassifier                |   0.512281 |            0.489598 |  0.489598 |   0.518924 |    0.0119965 |
     """
 
-    def __init__(self, verbose=0, ignore_warnings=True, custom_metric = None, predictions = False,random_state=42):
+    def __init__(
+        self,
+        verbose=0,
+        ignore_warnings=True,
+        custom_metric=None,
+        predictions=False,
+        random_state=42,
+        classifiers = "all"
+    ):
         self.verbose = verbose
         self.ignore_warnings = ignore_warnings
         self.custom_metric = custom_metric
         self.predictions = predictions
-        self.random_state =random_state
+        self.models = {}
+        self.random_state = random_state
+        self.classifiers = classifiers
 
     def fit(self, X_train, X_test, y_train, y_test):
         """Fit Classification algorithms to X_train and y_train, predict and score on X_test, y_test.
@@ -194,50 +260,69 @@ class LazyClassifier:
         names = []
         TIME = []
         predictions = {}
-        
-        if self.custom_metric != None:
+
+        if self.custom_metric is not None:
             CUSTOM_METRIC = []
-            
-        if type(X_train) is np.ndarray:
+
+        if isinstance(X_train, np.ndarray):
             X_train = pd.DataFrame(X_train)
             X_test = pd.DataFrame(X_test)
 
-        numeric_features = X_train.select_dtypes(
-            include=['int64', 'float64', 'int32', 'float32']).columns
-        categorical_features = X_train.select_dtypes(
-            include=['object']).columns
+        numeric_features = X_train.select_dtypes(include=[np.number]).columns
+        categorical_features = X_train.select_dtypes(include=["object"]).columns
+
+        categorical_low, categorical_high = get_card_split(
+            X_train, categorical_features
+        )
 
         preprocessor = ColumnTransformer(
             transformers=[
-                ('numeric', numeric_transformer, numeric_features),
-                ('categorical', categorical_transformer, categorical_features)
-            ])
+                ("numeric", numeric_transformer, numeric_features),
+                ("categorical_low", categorical_transformer_low, categorical_low),
+                ("categorical_high", categorical_transformer_high, categorical_high),
+            ]
+        )
 
-        for name, model in tqdm(CLASSIFIERS):
+        if self.classifiers == "all":
+            self.classifiers = CLASSIFIERS
+        else:
+            try:
+                temp_list = []
+                for classifier in self.classifiers:
+                    full_name = (classifier.__class__.__name__, classifier)
+                    temp_list.append(full_name)
+                self.classifiers = temp_list
+            except Exception as exception:
+                print(exception)
+                print("Invalid Classifier(s)")
+
+        for name, model in tqdm(self.classifiers):
             start = time.time()
             try:
-                if 'random_state' in model().get_params().keys():
-                    pipe = Pipeline(steps=[
-                        ('preprocessor', preprocessor),
-                        ('classifier', model(random_state = self.random_state))
-                    ])
+                if "random_state" in model().get_params().keys():
+                    pipe = Pipeline(
+                        steps=[
+                            ("preprocessor", preprocessor),
+                            ("classifier", model(random_state=self.random_state)),
+                        ]
+                    )
                 else:
-                    pipe = Pipeline(steps=[
-                        ('preprocessor', preprocessor),
-                        ('classifier', model())
-                    ])
+                    pipe = Pipeline(
+                        steps=[("preprocessor", preprocessor), ("classifier", model())]
+                    )
 
                 pipe.fit(X_train, y_train)
+                self.models[name] = pipe
                 y_pred = pipe.predict(X_test)
                 accuracy = accuracy_score(y_test, y_pred, normalize=True)
                 b_accuracy = balanced_accuracy_score(y_test, y_pred)
-                f1 = f1_score(y_test, y_pred, average='weighted')
+                f1 = f1_score(y_test, y_pred, average="weighted")
                 try:
                     roc_auc = roc_auc_score(y_test, y_pred)
                 except Exception as exception:
                     roc_auc = None
-                    if self.ignore_warnings == False:
-                        print("ROC AUC couldn't be calculated for "+name)
+                    if self.ignore_warnings is False:
+                        print("ROC AUC couldn't be calculated for " + name)
                         print(exception)
                 names.append(name)
                 Accuracy.append(accuracy)
@@ -245,52 +330,103 @@ class LazyClassifier:
                 ROC_AUC.append(roc_auc)
                 F1.append(f1)
                 TIME.append(time.time() - start)
-                if self.custom_metric != None:
+                if self.custom_metric is not None:
                     custom_metric = self.custom_metric(y_test, y_pred)
                     CUSTOM_METRIC.append(custom_metric)
                 if self.verbose > 0:
-                    if self.custom_metric != None:
-                        print({"Model": name,
-                               "Accuracy": accuracy,
-                               "Balanced Accuracy": b_accuracy,
-                               "ROC AUC": roc_auc,
-                               "F1 Score": f1,
-                               self.custom_metric.__name__: custom_metric,
-                              "Time taken": time.time() - start})
+                    if self.custom_metric is not None:
+                        print(
+                            {
+                                "Model": name,
+                                "Accuracy": accuracy,
+                                "Balanced Accuracy": b_accuracy,
+                                "ROC AUC": roc_auc,
+                                "F1 Score": f1,
+                                self.custom_metric.__name__: custom_metric,
+                                "Time taken": time.time() - start,
+                            }
+                        )
                     else:
-                        print({"Model": name,
-                               "Accuracy": accuracy,
-                               "Balanced Accuracy": b_accuracy,
-                               "ROC AUC": roc_auc,
-                               "F1 Score": f1,
-                              "Time taken": time.time() - start})
-                if self.predictions == True:
-                    predictions[name]=y_pred
+                        print(
+                            {
+                                "Model": name,
+                                "Accuracy": accuracy,
+                                "Balanced Accuracy": b_accuracy,
+                                "ROC AUC": roc_auc,
+                                "F1 Score": f1,
+                                "Time taken": time.time() - start,
+                            }
+                        )
+                if self.predictions:
+                    predictions[name] = y_pred
             except Exception as exception:
-                if self.ignore_warnings == False:
+                if self.ignore_warnings is False:
                     print(name + " model failed to execute")
                     print(exception)
-        if self.custom_metric == None:
-            scores = pd.DataFrame({"Model": names,
-                                   "Accuracy": Accuracy,
-                                   "Balanced Accuracy": B_Accuracy,
-                                   "ROC AUC": ROC_AUC,
-                                   "F1 Score": F1,
-                                   "Time Taken": TIME})
+        if self.custom_metric is None:
+            scores = pd.DataFrame(
+                {
+                    "Model": names,
+                    "Accuracy": Accuracy,
+                    "Balanced Accuracy": B_Accuracy,
+                    "ROC AUC": ROC_AUC,
+                    "F1 Score": F1,
+                    "Time Taken": TIME,
+                }
+            )
         else:
-            scores = pd.DataFrame({"Model": names,
-                                   "Accuracy": Accuracy,
-                                   "Balanced Accuracy": B_Accuracy,
-                                   "ROC AUC": ROC_AUC,
-                                   "F1 Score": F1,
-                                  self.custom_metric.__name__: CUSTOM_METRIC,
-                                  "Time Taken": TIME})
-        scores = scores.sort_values(
-            by='Balanced Accuracy', ascending=False).set_index('Model')
+            scores = pd.DataFrame(
+                {
+                    "Model": names,
+                    "Accuracy": Accuracy,
+                    "Balanced Accuracy": B_Accuracy,
+                    "ROC AUC": ROC_AUC,
+                    "F1 Score": F1,
+                    self.custom_metric.__name__: CUSTOM_METRIC,
+                    "Time Taken": TIME,
+                }
+            )
+        scores = scores.sort_values(by="Balanced Accuracy", ascending=False).set_index(
+            "Model"
+        )
 
-        if self.predictions == True:
+        if self.predictions:
             predictions_df = pd.DataFrame.from_dict(predictions)
-        return scores, predictions_df if self.predictions == True else scores
+        return scores, predictions_df if self.predictions is True else scores
+
+    def provide_models(self, X_train, X_test, y_train, y_test):
+        """
+        This function returns all the model objects trained in fit function.
+        If fit is not called already, then we call fit and then return the models.
+        Parameters
+        ----------
+        X_train : array-like,
+            Training vectors, where rows is the number of samples
+            and columns is the number of features.
+        X_test : array-like,
+            Testing vectors, where rows is the number of samples
+            and columns is the number of features.
+        y_train : array-like,
+            Training vectors, where rows is the number of samples
+            and columns is the number of features.
+        y_test : array-like,
+            Testing vectors, where rows is the number of samples
+            and columns is the number of features.
+        Returns
+        -------
+        models: dict-object,
+            Returns a dictionary with each model pipeline as value 
+            with key as name of models.
+        """
+        if len(self.models.keys()) == 0:
+            self.fit(X_train,X_test,y_train,y_test)
+
+        return self.models
+
+
+def adjusted_rsquared(r2, n, p):
+    return 1 - (1-r2) * ((n-1) / (n-p-1))
+
 
 # Helper class for performing classification
 
@@ -309,6 +445,8 @@ class LazyRegressor:
         When function is provided, models are evaluated based on the custom evaluation metric provided.
     prediction : bool, optional (default=False)
         When set to True, the predictions of all the models models are returned as dataframe.
+    regressors : list, optional (default="all")
+        When function is provided, trains the chosen regressor(s).
 
     Examples
     --------
@@ -316,60 +454,81 @@ class LazyRegressor:
     >>> from sklearn import datasets
     >>> from sklearn.utils import shuffle
     >>> import numpy as np
+
     >>> boston = datasets.load_boston()
     >>> X, y = shuffle(boston.data, boston.target, random_state=13)
     >>> X = X.astype(np.float32)
+
     >>> offset = int(X.shape[0] * 0.9)
     >>> X_train, y_train = X[:offset], y[:offset]
     >>> X_test, y_test = X[offset:], y[offset:]
-    >>> reg = LazyRegressor(verbose=0,ignore_warnings=False, custom_metric=None )
-    >>> models,predictions = reg.fit(X_train, X_test, y_train, y_test)
-    | Model                         |   R-Squared |     RMSE |   Time Taken |
-    |:------------------------------|------------:|---------:|-------------:|
-    | SVR                           |   0.877199  |  2.62054 |    0.0330021 |
-    | RandomForestRegressor         |   0.874429  |  2.64993 |    0.0659981 |
-    | ExtraTreesRegressor           |   0.867566  |  2.72138 |    0.0570002 |
-    | AdaBoostRegressor             |   0.865851  |  2.73895 |    0.144999  |
-    | NuSVR                         |   0.863712  |  2.7607  |    0.0340044 |
-    | GradientBoostingRegressor     |   0.858693  |  2.81107 |    0.13      |
-    | KNeighborsRegressor           |   0.826307  |  3.1166  |    0.0179954 |
-    | HistGradientBoostingRegressor |   0.810479  |  3.25551 |    0.820995  |
-    | BaggingRegressor              |   0.800056  |  3.34383 |    0.0579946 |
-    | MLPRegressor                  |   0.750536  |  3.73503 |    0.725997  |
-    | HuberRegressor                |   0.736973  |  3.83522 |    0.0370018 |
-    | LinearSVR                     |   0.71914   |  3.9631  |    0.0179989 |
-    | RidgeCV                       |   0.718402  |  3.9683  |    0.018003  |
-    | BayesianRidge                 |   0.718102  |  3.97041 |    0.0159984 |
-    | Ridge                         |   0.71765   |  3.9736  |    0.0149941 |
-    | LinearRegression              |   0.71753   |  3.97444 |    0.0190051 |
-    | TransformedTargetRegressor    |   0.71753   |  3.97444 |    0.012001  |
-    | LassoCV                       |   0.717337  |  3.9758  |    0.0960066 |
-    | ElasticNetCV                  |   0.717104  |  3.97744 |    0.0860076 |
-    | LassoLarsCV                   |   0.717045  |  3.97786 |    0.0490005 |
-    | LassoLarsIC                   |   0.716636  |  3.98073 |    0.0210001 |
-    | LarsCV                        |   0.715031  |  3.99199 |    0.0450008 |
-    | Lars                          |   0.715031  |  3.99199 |    0.0269964 |
-    | SGDRegressor                  |   0.714362  |  3.99667 |    0.0210009 |
-    | RANSACRegressor               |   0.707849  |  4.04198 |    0.111998  |
-    | ElasticNet                    |   0.690408  |  4.16088 |    0.0190012 |
-    | Lasso                         |   0.662141  |  4.34668 |    0.0180018 |
-    | OrthogonalMatchingPursuitCV   |   0.591632  |  4.77877 |    0.0180008 |
-    | ExtraTreeRegressor            |   0.583314  |  4.82719 |    0.0129974 |
-    | PassiveAggressiveRegressor    |   0.556668  |  4.97914 |    0.0150032 |
-    | GaussianProcessRegressor      |   0.428298  |  5.65425 |    0.0580051 |
-    | OrthogonalMatchingPursuit     |   0.379295  |  5.89159 |    0.0180039 |
-    | DecisionTreeRegressor         |   0.318767  |  6.17217 |    0.0230272 |
-    | DummyRegressor                |  -0.0215752 |  7.55832 |    0.0140116 |
-    | LassoLars                     |  -0.0215752 |  7.55832 |    0.0180008 |
-    | KernelRidge                   |  -8.24669   | 22.7396  |    0.0309792 |
+
+    >>> reg = LazyRegressor(verbose=0, ignore_warnings=False, custom_metric=None)
+    >>> models, predictions = reg.fit(X_train, X_test, y_train, y_test)
+    >>> model_dictionary = reg.provide_models(X_train, X_test, y_train, y_test)
+    >>> models
+    | Model                         | Adjusted R-Squared | R-Squared |  RMSE | Time Taken |
+    |:------------------------------|-------------------:|----------:|------:|-----------:|
+    | SVR                           |               0.83 |      0.88 |  2.62 |       0.01 |
+    | BaggingRegressor              |               0.83 |      0.88 |  2.63 |       0.03 |
+    | NuSVR                         |               0.82 |      0.86 |  2.76 |       0.03 |
+    | RandomForestRegressor         |               0.81 |      0.86 |  2.78 |       0.21 |
+    | XGBRegressor                  |               0.81 |      0.86 |  2.79 |       0.06 |
+    | GradientBoostingRegressor     |               0.81 |      0.86 |  2.84 |       0.11 |
+    | ExtraTreesRegressor           |               0.79 |      0.84 |  2.98 |       0.12 |
+    | AdaBoostRegressor             |               0.78 |      0.83 |  3.04 |       0.07 |
+    | HistGradientBoostingRegressor |               0.77 |      0.83 |  3.06 |       0.17 |
+    | PoissonRegressor              |               0.77 |      0.83 |  3.11 |       0.01 |
+    | LGBMRegressor                 |               0.77 |      0.83 |  3.11 |       0.07 |
+    | KNeighborsRegressor           |               0.77 |      0.83 |  3.12 |       0.01 |
+    | DecisionTreeRegressor         |               0.65 |      0.74 |  3.79 |       0.01 |
+    | MLPRegressor                  |               0.65 |      0.74 |  3.80 |       1.63 |
+    | HuberRegressor                |               0.64 |      0.74 |  3.84 |       0.01 |
+    | GammaRegressor                |               0.64 |      0.73 |  3.88 |       0.01 |
+    | LinearSVR                     |               0.62 |      0.72 |  3.96 |       0.01 |
+    | RidgeCV                       |               0.62 |      0.72 |  3.97 |       0.01 |
+    | BayesianRidge                 |               0.62 |      0.72 |  3.97 |       0.01 |
+    | Ridge                         |               0.62 |      0.72 |  3.97 |       0.01 |
+    | TransformedTargetRegressor    |               0.62 |      0.72 |  3.97 |       0.01 |
+    | LinearRegression              |               0.62 |      0.72 |  3.97 |       0.01 |
+    | ElasticNetCV                  |               0.62 |      0.72 |  3.98 |       0.04 |
+    | LassoCV                       |               0.62 |      0.72 |  3.98 |       0.06 |
+    | LassoLarsIC                   |               0.62 |      0.72 |  3.98 |       0.01 |
+    | LassoLarsCV                   |               0.62 |      0.72 |  3.98 |       0.02 |
+    | Lars                          |               0.61 |      0.72 |  3.99 |       0.01 |
+    | LarsCV                        |               0.61 |      0.71 |  4.02 |       0.04 |
+    | SGDRegressor                  |               0.60 |      0.70 |  4.07 |       0.01 |
+    | TweedieRegressor              |               0.59 |      0.70 |  4.12 |       0.01 |
+    | GeneralizedLinearRegressor    |               0.59 |      0.70 |  4.12 |       0.01 |
+    | ElasticNet                    |               0.58 |      0.69 |  4.16 |       0.01 |
+    | Lasso                         |               0.54 |      0.66 |  4.35 |       0.02 |
+    | RANSACRegressor               |               0.53 |      0.65 |  4.41 |       0.04 |
+    | OrthogonalMatchingPursuitCV   |               0.45 |      0.59 |  4.78 |       0.02 |
+    | PassiveAggressiveRegressor    |               0.37 |      0.54 |  5.09 |       0.01 |
+    | GaussianProcessRegressor      |               0.23 |      0.43 |  5.65 |       0.03 |
+    | OrthogonalMatchingPursuit     |               0.16 |      0.38 |  5.89 |       0.01 |
+    | ExtraTreeRegressor            |               0.08 |      0.32 |  6.17 |       0.01 |
+    | DummyRegressor                |              -0.38 |     -0.02 |  7.56 |       0.01 |
+    | LassoLars                     |              -0.38 |     -0.02 |  7.56 |       0.01 |
+    | KernelRidge                   |             -11.50 |     -8.25 | 22.74 |       0.01 |
     """
 
-    def __init__(self, verbose=0, ignore_warnings=True, custom_metric = None, predictions = False,random_state=42):
+    def __init__(
+        self,
+        verbose=0,
+        ignore_warnings=True,
+        custom_metric=None,
+        predictions=False,
+        random_state=42,
+        regressors="all",
+    ):
         self.verbose = verbose
         self.ignore_warnings = ignore_warnings
         self.custom_metric = custom_metric
         self.predictions = predictions
+        self.models = {}
         self.random_state = random_state
+        self.regressors = regressors 
 
     def fit(self, X_train, X_test, y_train, y_test):
         """Fit Regression algorithms to X_train and y_train, predict and score on X_test, y_test.
@@ -395,91 +554,148 @@ class LazyRegressor:
             Returns predictions of all the models in a Pandas DataFrame.
         """
         R2 = []
+        ADJR2 = []
         RMSE = []
         # WIN = []
         names = []
         TIME = []
         predictions = {}
-        
-        if self.custom_metric != None:
+
+        if self.custom_metric:
             CUSTOM_METRIC = []
 
-        if type(X_train) is np.ndarray:
+        if isinstance(X_train, np.ndarray):
             X_train = pd.DataFrame(X_train)
             X_test = pd.DataFrame(X_test)
 
-        numeric_features = X_train.select_dtypes(
-            include=['int64', 'float64', 'int32', 'float32']).columns
-        categorical_features = X_train.select_dtypes(
-            include=['object']).columns
+        numeric_features = X_train.select_dtypes(include=[np.number]).columns
+        categorical_features = X_train.select_dtypes(include=["object"]).columns
+
+        categorical_low, categorical_high = get_card_split(
+            X_train, categorical_features
+        )
 
         preprocessor = ColumnTransformer(
             transformers=[
-                ('numeric', numeric_transformer, numeric_features),
-                ('categorical', categorical_transformer, categorical_features)
-            ])
+                ("numeric", numeric_transformer, numeric_features),
+                ("categorical_low", categorical_transformer_low, categorical_low),
+                ("categorical_high", categorical_transformer_high, categorical_high),
+            ]
+        )
 
-        for name, model in tqdm(REGRESSORS):
+        if self.regressors == "all": 
+            self.regressors = REGRESSORS
+        else:
+            try:
+                temp_list = []
+                for regressor in self.regressors:
+                    full_name = (regressor.__class__.__name__, regressor)
+                    temp_list.append(full_name)
+                self.regressors = temp_list
+            except Exception as exception:
+                print(exception)
+                print("Invalid Regressor(s)")
+
+        for name, model in tqdm(self.regressors):
             start = time.time()
             try:
-                if 'random_state' in model().get_params().keys():
-                    pipe = Pipeline(steps=[
-                        ('preprocessor', preprocessor),
-                        ('regressor', model(random_state = self.random_state))
-                    ])
+                if "random_state" in model().get_params().keys():
+                    pipe = Pipeline(
+                        steps=[
+                            ("preprocessor", preprocessor),
+                            ("regressor", model(random_state=self.random_state)),
+                        ]
+                    )
                 else:
-                    pipe = Pipeline(steps=[
-                    ('preprocessor', preprocessor),
-                    ('regressor', model())
-                ])
+                    pipe = Pipeline(
+                        steps=[("preprocessor", preprocessor), ("regressor", model())]
+                    )
+
                 pipe.fit(X_train, y_train)
+                self.models[name] = pipe
                 y_pred = pipe.predict(X_test)
+
                 r_squared = r2_score(y_test, y_pred)
+                adj_rsquared = adjusted_rsquared(r_squared, X_test.shape[0], X_test.shape[1])
                 rmse = np.sqrt(mean_squared_error(y_test, y_pred))
+
                 names.append(name)
                 R2.append(r_squared)
+                ADJR2.append(adj_rsquared)
                 RMSE.append(rmse)
                 TIME.append(time.time() - start)
-                if self.custom_metric != None:
+
+                if self.custom_metric:
                     custom_metric = self.custom_metric(y_test, y_pred)
                     CUSTOM_METRIC.append(custom_metric)
 
                 if self.verbose > 0:
-                    if self.custom_metric != None:
-                        print({"Model": name,
-                               "R-Squared": r_squared,
-                               "RMSE": rmse,
-                               self.custom_metric.__name__: custom_metric,
-                              "Time taken": time.time() - start})
-                    else:
-                        print({"Model": name,
-                               "R-Squared": r_squared,
-                               "RMSE": rmse,
-                              "Time taken": time.time() - start})
-                if self.predictions == True:
-                    predictions[name]=y_pred
+                    scores_verbose = {
+                        "Model": name,
+                        "R-Squared": r_squared,
+                        "Adjusted R-Squared": adj_rsquared,
+                        "RMSE": rmse,
+                        "Time taken": time.time() - start,
+                    }
+
+                    if self.custom_metric:
+                        scores_verbose[self.custom_metric.__name__] = custom_metric
+
+                    print(scores_verbose)
+                if self.predictions:
+                    predictions[name] = y_pred
             except Exception as exception:
-                if self.ignore_warnings == False:
+                if self.ignore_warnings is False:
                     print(name + " model failed to execute")
                     print(exception)
-                    
-        if self.custom_metric == None:
-            scores = pd.DataFrame({"Model": names, 
-                                   "R-Squared": R2, 
-                                   "RMSE": RMSE,
-                                   "Time Taken": TIME})
-        else:
-            scores = pd.DataFrame({"Model": names, 
-                                   "R-Squared": R2, 
-                                   "RMSE": RMSE,
-                                  self.custom_metric.__name__: CUSTOM_METRIC,
-                                  "Time Taken": TIME})
-        scores = scores.sort_values(
-            by='R-Squared', ascending=False).set_index('Model')
-        
-        if self.predictions == True:
+
+        scores = {
+            "Model": names,
+            "Adjusted R-Squared": ADJR2,
+            "R-Squared": R2,
+            "RMSE": RMSE,
+            "Time Taken": TIME
+        }
+
+        if self.custom_metric:
+            scores[self.custom_metric.__name__] = CUSTOM_METRIC
+
+        scores = pd.DataFrame(scores)
+        scores = scores.sort_values(by="Adjusted R-Squared", ascending=False).set_index("Model")
+
+        if self.predictions:
             predictions_df = pd.DataFrame.from_dict(predictions)
-        return scores, predictions_df if self.predictions == True else scores
+        return scores, predictions_df if self.predictions is True else scores
+
+    def provide_models(self, X_train, X_test, y_train, y_test):
+        """
+        This function returns all the model objects trained in fit function.
+        If fit is not called already, then we call fit and then return the models.
+        Parameters
+        ----------
+        X_train : array-like,
+            Training vectors, where rows is the number of samples
+            and columns is the number of features.
+        X_test : array-like,
+            Testing vectors, where rows is the number of samples
+            and columns is the number of features.
+        y_train : array-like,
+            Training vectors, where rows is the number of samples
+            and columns is the number of features.
+        y_test : array-like,
+            Testing vectors, where rows is the number of samples
+            and columns is the number of features.
+        Returns
+        -------
+        models: dict-object,
+            Returns a dictionary with each model pipeline as value 
+            with key as name of models.
+        """
+        if len(self.models.keys()) == 0:
+            self.fit(X_train,X_test,y_train,y_test)
+
+        return self.models
+
 
 Regression = LazyRegressor
 Classification = LazyClassifier